(maint) explicitly include stdexcept
In fact_map.hpp and fact_resolver.hpp, exceptions defined in stdexcept
were being used but the header itself was not being excluded. In recent
versions of gcc and clang stdexcept was being included by another
library but in gcc 4.7 and clang 3.3 this was not included. This commit
explicitly requires stdexcept to avoid the version dependent behavior.