(maint) Fix submodule init and failure mode in Windows packaging
Without this change, packaging on Windows initializes the submodules on
the master branch, no matter what ref is requested.

Do submodule init and update after changing to the requested branch, to
ensure the correct submodules are initialized and at the right ref.

We are currently running into some errors where building native facter
will throw errors, but the script will continue to run. This isn't good,
especially because it means we ship an incomplete archive. This commit
is an attempt at having the native facter build script fail in a
consistent way when things go wrong.